ProALRS_55
developed and tested as of 2/27/2020, scripts run in ArcGIS Pro conda cloned environment.  This will migrate the existing Khub production transactional database to work with ArcGIS Pro ALRS.  The result is a pro ALRS enterprise geodatabase in SQL server, which then needs to be published to ArcGIS Portal compatible with PRO Alrs services.